moved into separate folders, updated library.json
